requirements.txt
pillow==2.5.1
